+++ src/expr/impl/src/lib.rs
@@ -34,7 +34,6 @@
 #![feature(result_flattening)]
 
 mod aggregate;
-mod external;
 mod scalar;
 mod table_function;
